+++ src/trainer/trainer.py
@@ -1236,7 +1236,10 @@ class Trainer:
         )
         y = torch.tensor(self.label_data.values.astype(np.float32), dtype=torch.float32)
 
-        D = [torch.tensor(value) for value in self.derived_data.values()]
+        D = [
+            torch.tensor(value.astype(np.float32))
+            for value in self.derived_data.values()
+        ]
         if src.setting["input_requires_grad"]:
             X.requires_grad = True
         dataset = Data.TensorDataset(X, *D, y)
